// ui/src/api/generated/models/doc.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.25.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DocSpec } from './doc-spec';
// May contain unused imports in some cases
// @ts-ignore
import type { Metadata } from './metadata';

/**
 * 
 * @export
 * @interface Doc
 */
export interface Doc {
    /**
     * 
     * @type {string}
     * @memberof Doc
     */
    'apiVersion': string;
    /**
     * 
     * @type {string}
     * @memberof Doc
     */
    'kind': string;
    /**
     * 
     * @type {Metadata}
     * @memberof Doc
     */
    'metadata': Metadata;
    /**
     * 
     * @type {DocSpec}
     * @memberof Doc
     */
    'spec': DocSpec;
}