// ui/src/api/generated/models/doc-library.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.25.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DocLibrarySpec } from './doc-library-spec';
// May contain unused imports in some cases
// @ts-ignore
import type { Metadata } from './metadata';

/**
 * 
 * @export
 * @interface DocLibrary
 */
export interface DocLibrary {
    /**
     * 
     * @type {string}
     * @memberof DocLibrary
     */
    'apiVersion': string;
    /**
     * 
     * @type {string}
     * @memberof DocLibrary
     */
    'kind': string;
    /**
     * 
     * @type {Metadata}
     * @memberof DocLibrary
     */
    'metadata': Metadata;
    /**
     * 
     * @type {DocLibrarySpec}
     * @memberof DocLibrary
     */
    'spec': DocLibrarySpec;
}